% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{add_rbf_features}
\alias{add_rbf_features}
\title{Add n_centroids RBF features to X}
\usage{
add_rbf_features(X, s, n_centroids, Xi = NULL)
}
\arguments{
\item{X}{covariate matrix}

\item{s}{median pairwise distance of points in X}

\item{n_centroids}{number of RBF features to add}

\item{Xi}{matrix of reference points to calculate RBF w.r.t}
}
\value{
X augmented covariate matrix
}
\description{
Add n_centroids RBF features to X
}
