% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\docType{class}
\name{AMS_data}
\alias{AMS_data}
\title{AMS data object class}
\description{
This reference class object is used to store the AMS metric of a classification
model at different decision thresholds. AMS is a performance measure which
includes the sample weightings and is defined by the Higgs Boson Kaggle Competition.
A vector of true sample classifications (0 or 1), a vector of estimated probabilities
from a model, and a vector of scaled sample weights are needed to initialise.
}
\section{Fields}{

\describe{
\item{\code{y}}{A vector of true sample classifications (0 or 1),}

\item{\code{prob}}{A vector of the samples estimated probabilities from a model}

\item{\code{weights}}{A vector of scaled sample weights.}

\item{\code{thresholds}}{A vector of 30 decision thresholds.}

\item{\code{ams}}{A vector of the AMS metric at each threshold.}
}}

\section{Methods}{

\describe{
\item{\code{calc_ams()}}{Calculate the AMS at each thresholds.}

\item{\code{initialize(y, prob, weights)}}{Provide true sample lables, estimated probabilities, and sample weights. A vector of descision thresholds is initalised.}

\item{\code{plot_ams()}}{Plot AMS against threshold.}
}}

