% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{ams_metric}
\alias{ams_metric}
\title{the AMS metric.
note s = sum_{i in B \cup G}w_i and b = sum_{i in B \cup G}w_i;
i.e. the sum of the weights of succesful signal classifications (TP)
and the sum of the weights of incorrect signal classifications (FP) respectively}
\usage{
ams_metric(s, b)
}
\arguments{
\item{s}{count of true positives}

\item{b}{count of false positives}
}
\description{
the AMS metric.
note s = sum_{i in B \cup G}w_i and b = sum_{i in B \cup G}w_i;
i.e. the sum of the weights of succesful signal classifications (TP)
and the sum of the weights of incorrect signal classifications (FP) respectively
}
