% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{ams_threshold}
\alias{ams_threshold}
\title{Calculate optimal threshold}
\usage{
ams_threshold(amss)
}
\arguments{
\item{amss}{a list of AMS_data objects}
}
\value{
optimal decision threshold considering all curves#
}
\description{
Calculate optimal threshold
}
