% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{average_auc}
\alias{average_auc}
\title{Calculate average AUC}
\usage{
average_auc(rocs)
}
\arguments{
\item{rocs}{a list of ROC_curve objects}
}
\value{
average AUC
}
\description{
Calculate average AUC
}
