% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{avg_median_pairwise_distance}
\alias{avg_median_pairwise_distance}
\title{Calculate average of median pairwise distances for between all adjacent points}
\usage{
avg_median_pairwise_distance(X)
}
\arguments{
\item{X}{covariate matrix}
}
\value{
s median pairwise distance
}
\description{
Calculate average of median pairwise distances for between all adjacent points
}
