% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_funcs.r
\name{backtrack_linesearch}
\alias{backtrack_linesearch}
\title{Backtrack linesearch}
\usage{
backtrack_linesearch(f, gradf, x, deltax, alpha, beta)
}
\arguments{
\item{f}{function to minimising}

\item{gradf}{gradient of f}

\item{x}{parameter to optimising over}

\item{deltax}{newton step}

\item{alpha}{linesearch parameter}

\item{beta}{linesearch update parameter}
}
\value{
stop to take
}
\description{
Backtracking linesearch to find "optimal" step size
}
