% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_funcs.r
\name{calc_K}
\alias{calc_K}
\title{Compute the kernel matrix over the (training) set X}
\usage{
calc_K(X, ckernel)
}
\arguments{
\item{X}{covariate matrix (nxd)}

\item{ckernel}{kernel function with args (x_i, x_j)}
}
\value{
K kernel matrix (nxn)
}
\description{
Compute the kernel matrix over the (training) set X
}
