% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/chillclimb.R
\name{chillclimb}
\alias{chillclimb}
\title{Compare Hill Climbing function}
\usage{
chillclimb(X, Co, y)
}
\arguments{
\item{X,Co,y}{This takes in the input of the vector for set of inputs, the randomly generated coefficients, and the mode of connection}
}
\description{
This function implements hill climbing algorithm by comparing various ascent methods and giving the result that has the optimal solution
}
\examples{
chillclimb(X,Co,y)
}
\keyword{ascent}

