% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{compute_AUC}
\alias{compute_AUC}
\title{Calculate the area under the ROC curve (AUC) as a metric of performance}
\usage{
compute_AUC(FP, TP)
}
\arguments{
\item{FP}{false positive rate (vector)}

\item{TP}{true positive rate (vector)}
}
\value{
AUC estimate (scalar)
}
\description{
Calculate the area under the ROC curve (AUC) as a metric of performance
}
