% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{compute_roc}
\alias{compute_roc}
\title{Compute receiver operating characteristic (ROC) curve}
\usage{
compute_roc(y, p_hat)
}
\arguments{
\item{y}{response vector}

\item{p_hat}{probability outputs from model}
}
\value{
list of false positive and true positive rates at different thresholds
}
\description{
Compute receiver operating characteristic (ROC) curve
}
