% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{count_b}
\alias{count_b}
\title{Count the number of false positives
need to make sure dims of y, y_hat and w are the same}
\usage{
count_b(y, y_hat, w)
}
\arguments{
\item{y}{response vector}

\item{y_hat}{predicted response vector}

\item{w}{weights}
}
\description{
Count the number of false positives
need to make sure dims of y, y_hat and w are the same
}
