% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{decide}
\alias{decide}
\title{Thresholding function}
\usage{
decide(p, thresh = 0.5)
}
\arguments{
\item{p}{vector of probabilities}

\item{thresh}{threshold over which we assign output of 1}
}
\description{
Thresholding function
}
