% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{generate_colours}
\alias{generate_colours}
\title{Generate distinct colours}
\usage{
generate_colours(ncolours)
}
\arguments{
\item{ncolours}{number of colours we want}
}
\value{
vector of hex colours
}
\description{
Generate distinct colours
}
