% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{get_const_features}
\alias{get_const_features}
\title{find colnames for columns that are constant (e.g. all 1, -999, NA etc)}
\usage{
get_const_features(X)
}
\arguments{
\item{X}{matrix of covariates}
}
\value{
list of column names
}
\description{
find colnames for columns that are constant (e.g. all 1, -999, NA etc)
}
