% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{get_model_idx}
\alias{get_model_idx}
\title{helper function to get the index corresponding to the model built on folds {l !=k} and for jet number j {1,2,3}, ordering columns by fold and then with nesting on j i.e. first six cols are k=1, j=1,2,3; k=2, j=1,2,3; etc}
\usage{
get_model_idx(j, k, K)
}
\arguments{
\item{j}{this jet group}

\item{k}{this fold}

\item{K}{number of folds}
}
\description{
helper function to get the index corresponding to the model built on folds {l !=k} and for jet number j {1,2,3}, ordering columns by fold and then with nesting on j i.e. first six cols are k=1, j=1,2,3; k=2, j=1,2,3; etc
}
