% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{get_rbf_centroids}
\alias{get_rbf_centroids}
\title{Get reference points for RBF centroids}
\usage{
get_rbf_centroids(X, n_centroids, idx = NULL)
}
\arguments{
\item{X}{covariate matrix}

\item{n_centroids}{number of RBF centroids}

\item{idx}{[Optional] location of RBF centroid reference points}
}
\value{
list of Xi (centroid points) and idx (location of them)
}
\description{
Get reference points for RBF centroids
}
