% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{get_subset_idx}
\alias{get_subset_idx}
\title{Select indices of rows in x which correspond to values of labels, which can be multiple elements}
\usage{
get_subset_idx(x, labels)
}
\arguments{
\item{x}{vector of labels}

\item{labels}{reference labels to compare to
e.g. x <- c("t", "b", "t", "v", "b"); labels <- c("t", "b")
get_subset_idx(x, labels) = T, T, T, F, T}
}
\description{
Select indices of rows in x which correspond to values of labels, which can be multiple elements
}
