% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{get_valid_cols}
\alias{get_valid_cols}
\title{Gets the columns from header that aren't in features_to_rm}
\usage{
get_valid_cols(header, features_to_rm, j)
}
\arguments{
\item{header}{list of column names e.g. names(X)}

\item{features_to_rm}{list of feature names we want to remove}

\item{j}{jet group}
}
\value{
index of columns to retain
}
\description{
Gets the columns from header that aren't in features_to_rm
}
