% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ghillclimb.R
\name{ghillclimb}
\alias{ghillclimb}
\title{Hill Climbing Graphing function}
\usage{
ghillclimb(X, Co, y, m)
}
\arguments{
\item{X,Co,y,m}{This takes in the input of the vector for set of inputs, the randomly generated coefficients, the mode of connection, and the hill climbing mode (1= Steepest, 2= Median, 3= Least)}
}
\description{
This function implements hill climbing algorithm with various ascent methods, accompanying a graph
}
\examples{
ghillclimb(X,Co,y,m)
}
\keyword{steepest}

