% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/hillclimb.R
\name{hillclimb}
\alias{hillclimb}
\title{Hill Climbing function}
\usage{
hillclimb(X, Co, y, m)
}
\arguments{
\item{X,Co,y,m}{This takes in the input of the vector for set of inputs, the randomly generated coefficients, the mode of connection, and the hill climbing mode (1= Steepest, 2= Median, 3= Least)}
}
\description{
This function implements hill climbing algorithm with various ascent methods specified the user
}
\examples{
hillclimb(X,Co,y,m)
}
\keyword{ascent}

