% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_funcs.r
\name{idx_higgs_mass}
\alias{idx_higgs_mass}
\title{Get boolean index for rows with missing/or not missing (depending on G/j) Higgs mass}
\usage{
idx_higgs_mass(X, j, G)
}
\arguments{
\item{X}{matrix of covariates}

\item{j}{jet group}

\item{G}{number of jet groups}
}
\value{
vector of bools
}
\description{
Get boolean index for rows with missing/or not missing (depending on G/j) Higgs mass
}
