% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_funcs.r
\name{idx_jet_cat}
\alias{idx_jet_cat}
\title{Get boolean vector of rows with j=0,1 or 2+}
\usage{
idx_jet_cat(nj, j)
}
\arguments{
\item{nj}{Vector of number of jets for each point}

\item{j}{jet group}
}
\description{
Get boolean vector of rows with j=0,1 or 2+
}
