% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{import_data}
\alias{import_data}
\title{Import data from the Higgs csv file. Reorganise into covariates, labels and supplementary data (weights, ids)
Replace -999s with NA
Standardize covariates}
\usage{
import_data(filepath = "atlas-higgs-challenge-2014-v2.csv")
}
\arguments{
\item{filepath}{str location of csv}
}
\value{
named list of X, y, w, kaggle_w, kaggle_s, e_id, nj
}
\description{
Import data from the Higgs csv file. Reorganise into covariates, labels and supplementary data (weights, ids)
Replace -999s with NA
Standardize covariates
}
