% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_funcs.r
\name{interior_point_fit}
\alias{interior_point_fit}
\title{Interior point fit}
\usage{
interior_point_fit(
  f,
  dualf,
  gradf,
  Hf,
  x,
  m,
  mu = 10,
  eps = 1e-06,
  eps_feas = 1e-06
)
}
\arguments{
\item{f}{objective function}

\item{dualf}{dual objective function}

\item{gradf}{residual vector}

\item{Hf}{Hessian for residual}

\item{x}{primal-dual point}

\item{m}{number of inequality constraints}

\item{mu}{interior-point step parameter}

\item{eps}{tolerance for problem}

\item{eps_feas}{tolerance for feasibility of primal-dual points}
}
\description{
x contains lambda (dual var) as well
}
