% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{inv_model_idx}
\alias{inv_model_idx}
\title{does the inverse procedure to get_model_idx, s.t. if idx <- get_model_idx(j, k, K)
then (j,k) <- inv_model_idx(idx) (if R output tuples)}
\usage{
inv_model_idx(idx, K)
}
\arguments{
\item{idx}{model index}

\item{K}{number of folds}
}
\value{
numeric pair of j and k
}
\description{
does the inverse procedure to get_model_idx, s.t. if idx <- get_model_idx(j, k, K)
then (j,k) <- inv_model_idx(idx) (if R output tuples)
}
