% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_funcs.r
\name{kernel_svm}
\alias{kernel_svm}
\title{Kernel SVM}
\usage{
kernel_svm(X, y, C, ckernel)
}
\arguments{
\item{X}{An nxd matrix with samples as rows and features as columns.}

\item{y}{A length-n vector of -1s and 1s indicating true sample classes.}

\item{C}{Regularisation parameter.}

\item{ckernel}{Kernel function with hyperparamters set}
}
\description{
Fit a kernel support vector machine for binary classification.
}
