% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_funcs.r
\name{l1_logistic_reg}
\alias{l1_logistic_reg}
\title{L1 constrained logistic regression}
\usage{
l1_logistic_reg(X, y, C = 1)
}
\arguments{
\item{X}{covariate matrix}

\item{y}{response vector}

\item{C}{[Optional] regularisation parameter, defaults to 1}
}
\value{
b vector of coefficients
}
\description{
L1 constrained logistic regression
}
