% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lehc_graph.R
\name{lehc_graph}
\alias{lehc_graph}
\title{Least Hill Climbing function - Graph}
\usage{
lehc_graph(X, Co, y)
}
\arguments{
\item{X,Co,y}{This takes in the input of the vector for set of inputs, the randomly generated coefficients, and the mode of connection}
}
\description{
This function implements hill climbing algorithm with Least ascent, generating a graph at the end of the result
}
\examples{
lehc_graph(X,Co,y)
}
\keyword{steepest}

