% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lhc_function.R
\name{lhc_function}
\alias{lhc_function}
\title{Least Hill Climbing function}
\usage{
lhc_function(X, Co)
}
\arguments{
\item{X,Co}{This takes in the input of the vector for set of inputs and the randomly generated coefficients}
}
\description{
This function implements hill climbing algorithm with least ascent
}
\examples{
lhc_function(X,Co)
}
\keyword{least}

