% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.r
\docType{class}
\name{logistic_model}
\alias{logistic_model}
\title{Logistic model object class}
\description{
This reference class object fits a binary classification model,
using the logistic_reg function. The model can be used to
predict the classes of new samples. The sample classes must be 0 or 1,
and the prediction returns the estimated probabilities that each sample is class 1.
A decision threshold should be subsequently.
}
\section{Fields}{

\describe{
\item{\code{X}}{An nxd matrix with samples as rows and features as columns.}

\item{\code{y}}{A length-n vector of 0s and 1s indicating true sample classes.}

\item{\code{coeffs}}{A length-d vector of model coefficients.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(X, y)}}{Provide X and y and the coeffs field will be calculated using logistic_reg}

\item{\code{predict(X_test)}}{Provide a matrix of new samples and a vector of P(y=1) is returned}
}}

