% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_funcs.r
\name{logistic_reg}
\alias{logistic_reg}
\title{Fit a logistic regression model by IRWLS - same thing glm does}
\usage{
logistic_reg(X, y, lambda = 0)
}
\arguments{
\item{X}{covariate matrix}

\item{y}{response vector}

\item{lambda}{[Optional] L2 regularisation parameter}

\item{r}{[Optional] weight vector}
}
\value{
b vector of coefficients
}
\description{
Fit a logistic regression model by IRWLS - same thing glm does
}
