% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mhc_function.R
\name{mhc_function}
\alias{mhc_function}
\title{Median Hill Climbing function}
\usage{
mhc_function(X, Co, y)
}
\arguments{
\item{X,Co,y}{This takes in the input of the vector for set of inputs, the randomly generated coefficients, and the mode of connection}
}
\description{
This function implements hill climbing algorithm with median ascent
}
\examples{
mhc_function(X,Co,y)
}
\keyword{steepest}

