% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/neighbor_function.R
\name{neighbor_function}
\alias{neighbor_function}
\title{Neighborhood function}
\usage{
neighbor_function(X, y)
}
\arguments{
\item{X}{This takes in the input of the vector for set of inputs, and the mode of connection (1st degree, 2nd degree, etc.)}
}
\description{
This function generates all possible neighbors for the vector X, given in a matrix form
}
\examples{
neighbor_function(X)
}
\keyword{neighborhood}

