% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_funcs.r
\name{newton_step}
\alias{newton_step}
\title{Compute newton step}
\usage{
newton_step(grad, H)
}
\arguments{
\item{grad}{gradient vector}

\item{H}{Hessian matrix}
}
\value{
step to take
}
\description{
Compute newton step
}
