% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{pairwise_distance}
\alias{pairwise_distance}
\title{Calculate distance for each row of X0 and X1}
\usage{
pairwise_distance(X0, X1)
}
\arguments{
\item{X0}{covariate matrix}

\item{X1}{covariate matrix}
}
\value{
vector of distances
}
\description{
Calculate distance for each row of X0 and X1
}
