% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{partition_data}
\alias{partition_data}
\title{Partition data into (random) folds for cross-validation.}
\usage{
partition_data(n, k, random = FALSE)
}
\arguments{
\item{n}{number of rows}

\item{k}{number of folds}

\item{random}{flag to choose whether to randomly select}
}
\value{
ind vector of integers denoting the OOS fold each row belongs to
Returns vector of indices denoting the OOS index, i_e. for rows with I_i=1, those are OOS for i=1
}
\description{
Partition data into (random) folds for cross-validation.
}
