% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{permute_matrix}
\alias{permute_matrix}
\title{Cyclic permutation of rows of X by r rows}
\usage{
permute_matrix(X, r = 1)
}
\arguments{
\item{X}{covariate matrix}

\item{r}{number of rows to permute (default=1)}
}
\description{
Cyclic permutation of rows of X by r rows
}
