% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{plot_amss}
\alias{plot_amss}
\title{define a function to plot multiple ams objects on the same axes}
\usage{
plot_amss(amss, title = NULL, min.max = TRUE, ...)
}
\arguments{
\item{amss}{list of ams objects}

\item{title}{str title to give plot, if null a default title is generated}
}
\description{
define a function to plot multiple ams objects on the same axes
}
