% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{plot_coefs}
\alias{plot_coefs}
\title{plots of parameter values by fold to compare and check consistency of models}
\usage{
plot_coefs(b)
}
\arguments{
\item{b}{matrix of coefficients (dxK)}
}
\description{
plots of parameter values by fold to compare and check consistency of models
}
