% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{plot_distributions}
\alias{plot_distributions}
\title{Density plots of variables}
\usage{
plot_distributions(X, variables = NULL, labels = NULL)
}
\arguments{
\item{X}{An nxd matrix with samples as rows and features as columns.}

\item{variables}{Optional vector of column names to be plotted}

\item{lables}{Optional vector of class labels to view distributions by class}
}
\description{
Given a matrix plot the density of the listed variables using ggplot2 facet_wrap.
}
\details{
If the matrix X is more than 10,000 samples, a random 10,000 samples will be selected
to keep the amount of data plotted reasonable.
}
