% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{plot_roc}
\alias{plot_roc}
\title{Plot ROC curve for particular model}
\usage{
plot_roc(y, p_hat, modelnum = "k j")
}
\arguments{
\item{y}{response vector}

\item{p_hat}{model output probabilities}
}
\description{
Plot ROC curve for particular model
}
