% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{plot_rocs}
\alias{plot_rocs}
\title{Plot Multiple ROC curves}
\usage{
plot_rocs(rocs, title = NULL, info = "", ...)
}
\arguments{
\item{rocs}{a list of ROC_curve objects}

\item{title}{str title to give plot}

\item{info}{additional info to add to the default title}
}
\description{
Plot Multiple ROC curves
}
