% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{poly_kernel}
\alias{poly_kernel}
\title{Define polynomial kernel}
\usage{
poly_kernel(x_i, x_j, b)
}
\arguments{
\item{x_i}{point in R^d}

\item{x_j}{point in R^d}

\item{b}{order of polynomial}
}
\value{
scalar of (1+x^Tx)^b
}
\description{
Define polynomial kernel
}
