% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{poly_transform}
\alias{poly_transform}
\title{Run polynomial transform on columns of X (of order b), removing output columns that are highly correlated}
\usage{
poly_transform(X, b = 2)
}
\arguments{
\item{X}{matrix of covariates}

\item{b}{order of polynomial}
}
\value{
X augmented matrix of covariates
}
\description{
Run polynomial transform on columns of X (of order b), removing output columns that are highly correlated
}
