% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{rbf_feature}
\alias{rbf_feature}
\title{Compute single RBF feature at some centroid i in idx (or xi in Xi)}
\usage{
rbf_feature(X, s, idx = NULL, xi = NULL)
}
\arguments{
\item{X}{covariate matrix}

\item{s}{median pairwise distance of points in X}

\item{idx}{[Optional] location of reference centroid}

\item{xi}{[Optional] reference centroid}
}
\description{
Compute single RBF feature at some centroid i in idx (or xi in Xi)
}
