% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{rbf_kernel}
\alias{rbf_kernel}
\title{RBF kernel}
\usage{
rbf_kernel(x_i, x_j, sigma)
}
\arguments{
\item{x_i}{point in R^d}

\item{x_j}{point in R^d}

\item{sigma}{bandwidth hyperparameter}
}
\description{
RBF kernel
}
