% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/rcoeff_function.R
\name{rcoeff_function}
\alias{rcoeff_function}
\title{Random Coefficient function}
\usage{
rcoeff_function(X)
}
\arguments{
\item{X}{This takes in the input of the vector for set of inputs}
}
\description{
This function generates a random set of constants
}
\examples{
rcoeff_function(X)
}
\keyword{coefficient}

