% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_specific_funcs.r
\name{reduce_features}
\alias{reduce_features}
\title{Reduce feature space dimensionality by exploiting redundancy}
\usage{
reduce_features(X)
}
\arguments{
\item{X}{matrix of covariates}
}
\value{
X augmented matrix of covariates
}
\description{
Reduce feature space dimensionality by exploiting redundancy
}
