% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{save_fig}
\alias{save_fig}
\title{function to wrap figure saving}
\usage{
save_fig(plot_func, filepath, filetype = pdf)
}
\arguments{
\item{plot_func}{partially called function of no arguments to generate plot}

\item{filepath}{string file path to save to}

\item{filetype}{type to save as (function name)}
}
\description{
function to wrap figure saving
}
