% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{scale_dat}
\alias{scale_dat}
\title{define a function to scale features of a matrix with reference to another matrix
useful because you can normalise X_train, and apply the same transformation to X_test
not designed for data with -999s!}
\usage{
scale_dat(X, ref, na.rm = FALSE)
}
\arguments{
\item{X}{matrix of covariates}

\item{ref}{matrix of covariates from which to calculate mu and sd}

\item{na.rm}{flag to be compatible with colMeans and sd (to ignore NA)}
}
\value{
augmented matrix of covariates, standardized and an intercept column
}
\description{
define a function to scale features of a matrix with reference to another matrix
useful because you can normalise X_train, and apply the same transformation to X_test
not designed for data with -999s!
}
