% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_specific_funcs.r
\name{set_features_to_rm}
\alias{set_features_to_rm}
\title{create list of feature names we want to omit based on jet group, or constant values/missing}
\usage{
set_features_to_rm(X, G, kI, nj)
}
\arguments{
\item{X}{matrix of covariates}

\item{G}{number of jet groups}

\item{kI}{fold indices (test label)}

\item{nj}{Vector of number of jets for each point}
}
\value{
nested list of column names
}
\description{
create list of feature names we want to omit based on jet group, or constant values/missing
}
