% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/shc_function.R
\name{shc_function}
\alias{shc_function}
\title{Steepest Hill Climbing function}
\usage{
shc_function(X, Co)
}
\arguments{
\item{X,Co}{This takes in the input of the vector for set of inputs and the randomly generated coefficients}
}
\description{
This function implements hill climbing algorithm with steepest ascent
}
\examples{
shc_function(X,Co)
}
\keyword{steepest}

