% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_funcs.r
\name{svm}
\alias{svm}
\title{Soft margin SVM}
\usage{
svm(X, y, C = 1)
}
\arguments{
\item{X}{An nxd matrix with samples as rows and features as columns.}

\item{y}{A length-n vector of -1s and 1s indicating true sample classes.}

\item{C}{Regularisation parameter.}
}
\description{
Fit a soft margin support vector machine for binary classification.
}
