% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{svm_predict}
\alias{svm_predict}
\title{Calculate predictions for test points by computing the kernel matrix over the training points and the kernel vector(/matrix) over test and training points}
\usage{
svm_predict(X_test, X_train, y, L, ckernel = lin_kernel)
}
\arguments{
\item{X_test}{matrix of covariate test points}

\item{X_train}{matrix of covariate training points}

\item{y}{response vector}

\item{L}{?}

\item{ckernel}{kernel function k(x_i,x_j)}
}
\description{
Calculate predictions for test points by computing the kernel matrix over the training points and the kernel vector(/matrix) over test and training points
}
