% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/thillclimb.R
\name{thillclimb}
\alias{thillclimb}
\title{Table Hill Climbing function}
\usage{
thillclimb(y, No, Ev, St)
}
\arguments{
\item{y,No,Ev,St}{This takes in the mode of connection, number of inputs, expected value of each input, and the number of sets generated}
}
\description{
This function implements hill climbing algorithm by comparing various ascent methods and giving the tables of solution
}
\examples{
thillclimb(y,No,Ev,St)
}
\keyword{ascent}

