% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_funcs.r
\name{trig_kernel}
\alias{trig_kernel}
\title{Define trigonometric kernel}
\usage{
trig_kernel(x_i, x_j, b = 0)
}
\arguments{
\item{x_i}{point in R^d}

\item{x_j}{point in R^d}

\item{b}{order of polynomial}
}
\description{
Define trigonometric kernel
}
