% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_funcs.r
\name{tuned_kernel}
\alias{tuned_kernel}
\title{Function factory to partially call kernel function to return the kernel function with it's hyperparameters set}
\usage{
tuned_kernel(ckernel, ...)
}
\arguments{
\item{ckernel}{kernel function with args (x_i,x_j,hyper)}
}
\value{
function with args (x_i,x_j)
}
\description{
Function factory to partially call kernel function to return the kernel function with it's hyperparameters set
}
