% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{ams_metric}
\alias{ams_metric}
\title{Calculate AMS metric}
\usage{
ams_metric(s, b)
}
\arguments{
\item{s}{count of true positives}

\item{b}{count of false positives}
}
\description{
\eqn{s = sum_{i in B \cup G}w_i}
\eqn{b = sum_{i in B \cup G}w_i};
i.e. s is the sum of the weights of successful signal classifications (TP)
and b is the sum of the weights of incorrect signal classifications (FP)
}
