% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_funcs.r
\name{backtrack_linesearch}
\alias{backtrack_linesearch}
\title{backtracking linesearch to find "optimal" step size}
\usage{
backtrack_linesearch(f, gradf, x, deltax, alpha, beta)
}
\arguments{
\item{f}{function we're minimising}

\item{gradf}{gradient of f}

\item{x}{parameter we're optimising over}

\item{deltax}{newton step}

\item{alpha}{linesearch parameter}

\item{beta}{linesearch update parameter}
}
\description{
backtracking linesearch to find "optimal" step size
}
