% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{calculate_ams_partition}
\alias{calculate_ams_partition}
\title{Calculate the AMS}
\usage{
calculate_ams_partition(y, y_hat, w, sum_w = NULL)
}
\arguments{
\item{y}{response vector}

\item{y_hat}{predicted response vector}

\item{w}{weights}

\item{sum_w}{total sum of weights for renormalisation}
}
\value{
ams
}
\description{
Calculate the AMS
}
