% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{import_data}
\alias{import_data}
\title{Import raw LHC data}
\usage{
import_data(filepath = "atlas-higgs-challenge-2014-v2.csv")
}
\arguments{
\item{filepath}{str location of csv}
}
\value{
named list of X, y, w, kaggle_w, kaggle_s, e_id, nj
}
\description{
This function provides a standard way to load in the LHC dataset for our analysis pipeline.
Undefined data (-999s) are replaces with NAs, columns are split into variables,
labels and supplementary data.
}
