% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lhc_function.R
\name{lhc_function}
\alias{lhc_function}
\title{Least Hill Climbing function}
\usage{
lhc_function(X, Co, y)
}
\arguments{
\item{X,Co,y}{This takes in the input of the vector for set of inputs, the randomly generated coefficients, and the mode of connection}
}
\description{
This function implements hill climbing algorithm with least ascent
}
\examples{
lhc_function(X,Co,y)
}
\keyword{steepest}

