% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_funcs.r
\name{logistic_reg}
\alias{logistic_reg}
\title{Logistic regression}
\usage{
logistic_reg(X, y, lambda = 0)
}
\arguments{
\item{X}{covariate matrix}

\item{y}{response vector}

\item{lambda}{[Optional] L2 regularisation parameter, defaults to 0.}

\item{r}{[Optional] weight vector}
}
\value{
b vector of coefficients
}
\description{
Fit a logistic regression model by IRWLS.
This is the same method that glm uses when family="binomial".
}
