% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mhc_function.R
\name{mhc_function}
\alias{mhc_function}
\title{Median Hill Climbing function}
\usage{
mhc_function(X, Co)
}
\arguments{
\item{X,Co}{This takes in the input of the vector for set of inputs and the randomly generated coefficients}
}
\description{
This function implements hill climbing algorithm with median ascent
}
\examples{
mhc_function(X,Co)
}
\keyword{median}

