% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.r
\name{plot_rocs}
\alias{plot_rocs}
\title{Compute receiver operating characteristic (ROC) curve}
\usage{
plot_rocs(rocs, title = "ROC curves")
}
\arguments{
\item{FP}{false positive rate (vector)}

\item{TP}{true positive rate (vector)}

\item{y}{response vector}

\item{p_hat}{model output probabilities
plots of parameter values by fold to compare and check consistency of models}

\item{b}{matrix of coefficients (dxK)}
}
\value{
list of false positive and true positive rates at different thresholds
Calculate the area under the ROC curve (AUC) as a metric of performance

AUC estimate (scalar)
Plot ROC curve for particular model
}
\description{
Compute receiver operating characteristic (ROC) curve
}
