% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funcs.r
\name{poly_transform}
\alias{poly_transform}
\title{Polynomial transform}
\usage{
poly_transform(X, b = 2)
}
\arguments{
\item{X}{matrix of covariates}

\item{b}{order of polynomial}
}
\value{
augmented matrix of covariates
}
\description{
Run a b-degree polynomial transform on the columns of X (of order b)
}
