% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/shc_graph.R
\name{shc_graph}
\alias{shc_graph}
\title{Steepest Hill Climbing function - Graph}
\usage{
shc_graph(X, Co)
}
\arguments{
\item{X,Co}{This takes in the input of the vector for set of inputs and the randomly generated coefficients}
}
\description{
This function implements hill climbing algorithm with steepest ascent, generating a graph at the end of the result
}
\examples{
shc_function(X,Co)
}
\keyword{steepest}

